src/api-service/__app__/requirements.txt
azure-cli-core~=2.10.0
azure-cli-nspkg==3.0.4
azure-common~=1.1.25
azure-core~=1.7.0
azure-cosmosdb-nspkg==2.0.2
azure-cosmosdb-table==1.0.6
azure-devops==6.0.0b2
azure-functions~=1.3.0
azure-graphrbac~=0.61.1
azure-identity~=1.3.1
azure-keyvault-keys~=4.1.0
azure-keyvault-secrets~=4.1.0
azure-keyvault~=4.1.0
azure-mgmt-compute~=13.0.0
azure-mgmt-core~=1.0.0
azure-mgmt-cosmosdb~=0.16.0
azure-mgmt-loganalytics~=0.7.0
azure-mgmt-network~=11.0.0
azure-mgmt-resource~=10.1.0
azure-mgmt-storage~=11.1.0
azure-mgmt-subscription~=0.6.0
azure-nspkg==3.0.2
azure-servicebus==0.50.3
azure-storage-blob==2.0.1
azure-storage-common==2.1.0
azure-storage-queue==12.1.2
jinja2~=2.11.2
msrestazure~=0.6.3
opencensus-ext-azure~=1.0.2
pydantic~=1.6.1
PyJWT~=1.7.1
requests~=2.24.0
memoization~=0.3.1
# onefuzztypes version is set during build
onefuzztypes==0.0.0
